elfobj: ELFOBJ-20 assorted 32-bit overflow / narrowing hardening
- comdat_set_add grew its item array with a raw realloc(items, ncap *
  sizeof) multiply; use the overflow-checked/capped elf__reallocarray.
- The DWARF line parser narrowed a 64-bit-DWARF unit_length into uint32_t,
  corrupting unit sizing for > 4 GiB sections on a 64-bit build; keep it
  size_t.
- The AArch64/x86 GNU-property validators computed the loop guard and
  decrement as 8 + pr_datasz / 8 + sz in 32-bit unsigned arithmetic, which
  wraps for an attacker datasz near 4 GiB; compute in uint64.
- The ARM/RISC-V attribute-section walkers computed end = start + sect_len
  and checked end > size afterwards, wrapping on the 32-bit target; bound
  via sect_len > size - start before adding.

All are low-impact (most need ~4 GiB payloads) hardening.

Verified: gnu_ext/debug_unwind/link tests pass, the DWARF-line and
ARM-attribute AddressSanitizer/UBSan repros stay clean, native and target
builds pass. (test_validate_hardening and test_multiarch_api fail
identically on the audit-base commit - pre-existing, not from this audit.)

Finding ELFOBJ-20 in docs/elfobj-audit-2026-07.md.

+++ usr.lib/elfobj/src/elf_dwarf_line.c
@@ -57,7 +57,7 @@ elf_err_t elf_dwarf_get_line_info(elfobj_t *obj, uint64_t addr, char **out_file,
 
     while (off < size) {
         if (off + 4 > size) break;
-        uint32_t unit_length = elf__rd32(data + off, obj->endian);
+        size_t unit_length = elf__rd32(data + off, obj->endian);
         off += 4;
 
         int is_64bit = 0;
@@ -67,7 +67,7 @@ elf_err_t elf_dwarf_get_line_info(elfobj_t *obj, uint64_t addr, char **out_file,
             uint64_t len64 = elf__rd64(data + off, obj->endian);
             off += 8;
             if (len64 > size - off) break;
-            unit_length = (uint32_t)len64;
+            unit_length = (size_t)len64;
         }
 
         /* off + unit_length can wrap size_t on the 32-bit target; compare
